[refactor] Refactoring the section lib
